third_party/WebKit/LayoutTests/http/tests/serviceworker/controller-on-reload.html

<!DOCTYPE html>
<title>Service Worker: Controller on reload</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var t = async_test('controller is set upon reload after registration');
t.step(function() {
    var reloaded = false;
    var scope = 'resources/blank.html';
    navigator.serviceWorker.unregister(scope).then(
        doTest,
        unreached_rejection(t, 'Unregister should not fail')
    );

    function doTest() {
        with_iframe(scope, t.step_func(onIframeLoad));
    }

    function onIframeLoad(frame) {
        var w = frame.contentWindow;
        if (reloaded) {
            assert_true(w.navigator.serviceWorker.controller instanceof w.ServiceWorker,
                        'controller should be a ServiceWorker object upon reload');
            service_worker_unregister_and_done(t, scope);
            return;
        }

        w.navigator.serviceWorker.register('worker-no-op.js',
                                           {scope: 'blank.html'})
        .then(function(registration) {
            return wait_for_update(t, registration);
          })
        .then(function(worker) {
            return wait_for_state(t, worker, 'activated');
          })
        .then(function() {
            assert_equals(w.navigator.serviceWorker.controller, null,
                          'controller should be null until the document is reloaded');
            reloaded = true;
            w.location.reload();
          })
        .catch(unreached_rejection(t));
    }
});
</script>
</body>
